4.23

+++ controllers/blog.js
@@ -1,5 +1,7 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog.js");
+const User = require("../models/user");
+const jwt = require("jsonwebtoken");
 
 blogsRouter.get("/", async (request, response) => {
   try {
@@ -13,13 +15,37 @@ blogsRouter.get("/", async (request, response) => {
 
 blogsRouter.post("/", async (request, response) => {
   try {
-    const { title, url } = request.body;
-
+    const { title, url, author, likes } = request.body;
+    const user = request.user;
     if (!title || !url) {
       return response.status(400).json({ error: "title or url missing" });
     }
-    const blog = new Blog(request.body);
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: "token invalid" });
+    }
+    const userDb = await User.findById(decodedToken.id);
+
+    console.log(userDb);
+    if (!userDb) {
+      return response.status(400).json({ error: "invalid user ID" });
+    }
+
+    const blog = new Blog({
+      title,
+      url,
+      likes: likes || 0,
+      author,
+      user: userDb.id,
+    });
+
     const result = await blog.save();
+
+    userDb.blogs = userDb.blogs.concat(result.id);
+    await userDb.save();
+
     response.status(201).json(result);
   } catch (error) {
     console.error(error);
@@ -29,12 +55,34 @@ blogsRouter.post("/", async (request, response) => {
 
 blogsRouter.delete("/:id", async (request, response) => {
   try {
-    const result = await Blog.findByIdAndRemove(request.params.id);
-    if (result) {
-      response.status(204).end();
-    } else {
-      response.status(404).end();
+    const { token, user } = request;
+    if (!token) {
+      return response.status(401).json({ error: "token missing or invalid" });
     }
+
+    const decodedToken = jwt.verify(token, process.env.SECRET); // Use your secret key
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: "token invalid" });
+    }
+
+    const blog = await Blog.findById(request.params.id);
+    if (!blog) {
+      return response.status(404).json({ error: "blog not found" });
+    }
+
+    const userDb = await User.findById(decodedToken.id);
+    if (!user) {
+      return response.status(401).json({ error: "user not found" });
+    }
+
+    if (blog.user.toString() !== userDb.id.toString()) {
+      return response
+        .status(403)
+        .json({ error: "not authorized to delete this blog" });
+    }
+
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
   } catch (error) {
     console.error(error);
     response.status(500).send("Internal Server Error");
@@ -46,10 +94,15 @@ blogsRouter.put("/:id", async (request, response) => {
   const blog = {
     likes,
   };
-  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {
-    new: true,
-  });
-  response.json(updatedBlog);
+  try {
+    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {
+      new: true,
+    });
+    response.json(updatedBlog);
+  } catch (error) {
+    console.error(error);
+    response.status(500).send("Internal Server Error");
+  }
 });
 
 module.exports = blogsRouter;
